benchmarks/inference/mii/requirements.txt
transformers
matplotlib
deepspeed-mii>=0.2.0
vllm>=0.2.7
numpy